pyproject.toml
[tool.poetry]
name = "coffee shop"
version = "0.1.0"
description = "coffee shop api"
authors = ["Kambarov Damir"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
django = "~4.2"
djangorestframework = "^3.15.2"
psycopg = {version = "^3.2.6", extras = ["binary"]}
djangorestframework-simplejwt = "^5.5.0"
python-dotenv = "^1.1.0"
drf-spectacular = "^0.28.0"
django-filter = "^25.1"
dj-rest-auth = "^7.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
isort = "^6.0.0"
wemake-python-styleguide = "0.18.0"
flake8 = "^7.0.0"
flake8-html = "^0.4.3"
autoflake = "^2.3.1"
pyproject-flake8 = "^7.0.0"
black = "^25.1.0"
django-debug-toolbar = "^5.0.1"

[tool.black]
line-length = 120
target-version = ["py312"]
exclude = """
(
    .venv
    | flake-report
    | venv
    | migrations
    | tests
)
"""

[tool.isort]
profile = "wemake"
default_section = "THIRDPARTY"
known_first_party = ["config", "users", "base", "api", "shop"]
skip_glob = ["**/__pycache__", ".venv", "flake-report", "venv", "**/migrations", "**/tests"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
line_length = 120

[tool.autoflake]
imports = ["django", "djangorestframework"]

[tool.flake8]
inline-quotes = "\""
format = "wemake"
show-source = true
statistics = "false"
doctests = true
select = "C,E,F,W,B,B950"
ignore = """
C812,C815,C816,C408,C901,\
W503,\
WPS100,WPS110,WPS114,WPS115,\
WPS201,WPS202,WPS203,WPS210,WPS211,WPS213,WPS214,WPS215,WPS216,WPS219,WPS220,WPS221,WPS222,WPS226,WPS230,WPS231,WPS234,WPS235,WPS237,\
WPS305,WPS306,WPS318,WPS319,WPS320,WPS323,WPS332,WPS337,WPS347,WPS348,\
WPS400,WPS402,WPS404,WPS407,WPS410,WPS412,WPS421,WPS425,WPS430,WPS431,WPS432,WPS442,WPS462,WPS463,\
WPS501,WPS503,WPS510,WPS529,\
WPS601,WPS602,WPS615
"""
max-try-body-length = 5
max-line-length = 120
strictness = "long"
docstring-style = "numpy"
exclude = ["**/__pycache__", ".venv", "**/tests/**", "flake-report", "venv", "**/migrations/**", "app/manage.py"]
min-name-length = 1
per-file-ignores = [
    "app/manage.py:WPS326,WPS433",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
